Resolve "Challenge mapper"

Closes #41

See merge request scrum_team_3/idatt2106_2024_03_backend!20

+++ src/main/java/idatt2106/systemutvikling/sparesti/dto/ChallengeDTO.java
@@ -28,7 +28,7 @@ public class ChallengeDTO {
 
   private LocalDateTime expirationDate;
 
-  private Long recurring;
+  private int recurring;
 
   private boolean isActive;
 }
